network_monitoring/requirements.txt
scapy==2.5.0
flask==3.0.0
flask-cors==4.0.0
google-generativeai==0.3.2
firebase-admin==6.3.0
pyshark==0.4.6
ipaddress==1.0.23
werkzeug==3.0.1
requests==2.31.0